+++ notification-domain/src/main/kotlin/io/github/v1servicenotification/detail/spi/PostDetailUserSpi.kt
@@ -1,4 +1,4 @@
-package io.github.v1servicenotification.detail.postDetail.spi
+package io.github.v1servicenotification.detail.spi
 
 import java.util.UUID
