+++ notification-domain/src/main/kotlin/io/github/v1servicenotification/detail/spi/QueryDetailRepositorySpi.kt
@@ -1,7 +1,7 @@
-package io.github.v1servicenotification.detail.queryDetail.spi
+package io.github.v1servicenotification.detail.spi
 
 import io.github.v1servicenotification.annotation.Spi
-import io.github.v1servicenotification.detail.queryDetail.spi.dto.DetailModel
+import io.github.v1servicenotification.detail.spi.dto.DetailModel
 import java.util.*
 
 @Spi
